Rename BaseLayout to DefaultLayout for clarity
